Refactor logging setup and remove redundant file
